question/0105 Construct Binary Tree from Preorder and Inorder Traversal_test.go
package question

import (
	"reflect"
	"testing"
)

func Test_buildTree105(t *testing.T) {
	type args struct {
		preorder []int
		inorder  []int
	}
	tests := []struct {
		name string
		args args
		want *TreeNode
	}{
		{
			name: "Test1",
			args: args{
				preorder: []int{3, 9, 20, 15, 7},
				inorder:  []int{9, 3, 15, 20, 7},
			},
			want: makeTree([]int{3, 9, 20, -1, -1, 15, 7}, -1),
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := buildTree105(tt.args.preorder, tt.args.inorder); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("buildTree105() = %v, want %v", got, tt.want)
			}
		})
	}
}
